Crate nd_util

Structs

ArcAnyhowError

An Arc’ed anyhow error.

DropRemovePath

Asyncronously remove a file at a path on drop.

Functions

download_to_file

Download a url using a GET request to a tokio file.

download_to_path

Using the given client, download the file at a url to a given path.

push_extension

Push an extension to a PathBuf.

try_create_dir

Try to create a dir at the given path.

try_remove_dir

Try to remove a dir at the given path.

with_push_extension

Push an extension to a Path, returning a new PathBuf.